Moved claim_verfier and added a unit test for string scope
